big ui update and front end change

+++ cfcapp/models.py
@@ -9,7 +9,7 @@ class SalesFormData(models.Model):
     sales_rep = models.ForeignKey(User, on_delete=models.SET_NULL, blank=True, null=True)
     company_name = models.CharField(max_length=255, blank=False)
     contact_name = models.CharField(max_length=255, blank=False)
-    total_spent = models.FloatField(blank=False)
+    total_spent = models.IntegerField(blank=False)
     address = models.CharField(max_length=255, blank=False)
     secondary_address = models.CharField(max_length=255, blank=True)
     city = models.CharField(max_length=255, blank=False)
